src/backend/requirements.txt
# Azure Translator Solution Accelerator - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Azure SDKs
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-storage-queue==12.9.0
azure-data-tables==12.4.4
azure-monitor-opentelemetry==1.2.0
openai==1.12.0

# Application Insights
opencensus-ext-azure==1.1.13
opencensus-ext-flask==0.8.0
applicationinsights==0.11.10

# CORS
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
backoff==2.2.1

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Validation
email-validator==2.1.0

